add login, logout, authentication

+++ base_app/views.py
@@ -1,9 +1,44 @@
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
+
 from .forms import RoomForm
 from .models import Room, Topic
 
 # Create your views here.
+
+
+def loginPage(request):
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except Exception:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or Password are incorrect...')
+
+    context = {
+
+    }
+    return render(request, 'base_app/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -33,6 +68,7 @@ def room(request, pk):
     return render(request, 'base_app/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
 
@@ -48,9 +84,13 @@ def createRoom(request):
     return render(request, 'base_app/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -64,8 +104,12 @@ def updateRoom(request, pk):
     return render(request, 'base_app/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
 
     if request.method == 'POST':
         room.delete()
